Use raw pointer in static_dict! macro (rust static_mut_refs lint)
This lint will become deny by default in Rust 2024 edition: https://doc.rust-lang.org/edition-guide/rust-2024/static-mut-references.html

As for the solution, https://users.rust-lang.org/t/whats-the-correct-way-of-doing-static-mut-in-2024-rust/120403/17 for helpful for me.
It makes sense to use a raw pointer rather than a reference when the only purpose is to pass the pointer to a C method.